Edit the message request endpoint for chat domain - not require destination variable anymore
