+++ .gitignore
@@ -36,3 +36,4 @@ build/
 logs/
 mock_data/sql
 container/data/
+container/prepare/sql
