+++ src/main/java/com/vivacon/dto/request/AddParticipantMessage.java
@@ -1,0 +1,14 @@
+package com.vivacon.dto.request;
+
+public class AddParticipantMessage extends MessageRequest{
+
+    private String username;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+}
