+++ src/main/java/com/vivacon/service/impl/ConversationServiceImpl.java
@@ -109,7 +109,7 @@ public class ConversationServiceImpl implements ConversationService {
     @Override
     public PageDTO<OutlineConversation> findAllByCurrentAccount(Optional<String> order, Optional<String> sort, Optional<Integer> pageSize, Optional<Integer> pageIndex) {
         Pageable pageable = PageableBuilder.buildPage(order, sort, pageSize.isPresent() ? pageSize : Optional.of(Integer.MAX_VALUE), pageIndex, Conversation.class);
-        Page<Conversation> conversations = conversationRepository.findAllIdByPrincipalUsername(accountService.getCurrentAccount().getUsername(), pageable);
+        Page<Conversation> conversations = conversationRepository.findAllByPrincipalUsername(accountService.getCurrentAccount().getUsername(), pageable);
         return PageMapper.toPageDTO(conversations, conversation -> conversationMapper.toOutlineConversation(conversation));
     }
